+++ msd_pytorch/tests/test_msd_model.py
@@ -1,7 +1,6 @@
 import torch
 from msd_pytorch.msd_model import MSDModel
 from torch.utils.data import TensorDataset, DataLoader
-from pytest import approx
 
 
 def test_normalization():
@@ -15,26 +14,22 @@ def test_normalization():
             # Network
             width = 1
             depth = 10
-            sample_size, num_channels, shape = 1, 1, (1000, 1000)
+            sample_size, num_channels, shape = 1, 3, (100, 100)
             model = MSDModel(num_channels, num_channels, depth, width)
 
             # Data
+            # We make the input and target dataset equal, which makes
+            # their mean and std equal. This should ensure that the
+            # scale_in and scale_out module are each others' inverse.
             d_in = mean + torch.randn(sample_size, num_channels, *shape) * std
-            d_out = mean + torch.randn(sample_size, num_channels, *shape) * std
+            d_out = d_in.clone()
             ds = TensorDataset(d_in, d_out)
             dl = DataLoader(ds, sample_size)
 
             # Calculate normalization:
             model.set_normalization(dl)
 
-            (input, _), *_ = dl
-
-            # Check input layer scaling
-            l0 = model.scale_in(input)
-            assert l0.mean().item() == approx(0, abs=1e-4, rel=1e-4)
-            assert l0.std().item() == approx(1.0, rel=1e-4)
-
-            # Check output layer scaling
-            l1 = model.scale_out(l0)
-            assert l1.mean().item() == approx(mean, abs=1e-2, rel=1e-2)
-            assert l1.std().item() == approx(std, rel=1e-2)
+            # Ensure that scale_out is the inverse of scale_in:
+            x = torch.randn_like(d_in)
+            y = model.scale_out(model.scale_in(x))
+            assert abs(x - y).mean() < 1e-4